First implementation
This is the very first and prototype implementation of the app. As it
was developed in a speedrun, things are dirty, code is not well
organized and technologies are mixed together in the wrong way. So, a
deep cleanup needs to be done, along with best practices implementation
and removal of unuseful classes. Stay tuned!

Signed-off-by: Matteo Principe <matteo.principe@keyless.io>