First working version of local/remote datastore
This commit introduces a first version of the repository working both
with local and remote data. Local and remote data sources are merged in
order to upload local whenever the remote responds, and display saved
data when the remote is not available. Still, code needs to be cleaned
up as I needed to learn a lot of stuff while developing

Signed-off-by: Matteo Principe <matteo.principe@keyless.io>