Better data sources logic
The implementation of the data sources was changed: we now use suspend
functions in order to retrieve data from local db or network, rather
than creating flows, merging them and so on (which was far more
complicated). We also correctly store data on local db when we
successfully retrieve it from the internet.

Signed-off-by: Matteo Principe <matteo.principe@keyless.io>