Implement filtering API, standard repo contract
This commit introduces the possibility to issue an API call with enabled
filtering. This API is already able to use all filtering, altough the UI
is't ready (and of course location is missing, so no distance filtering
for now). Moreover, now all the BreweryRepositoryImpl methods follow a
standard contract: they all return flows as they need to be collected
asynchronously by the viewmodel.